Tidy list of assets
